RIG - check forbidden values and pass length
Allow the Generator to receive an array of forbidden vars on init
which are checked along with the hash of existing vars prior to
saving it in the hash. This should prevent the unlikely case of,
for argument's sake, a PSH RIG producing "true" which would be
"$true" after template interpolation.

Permit :get (aka :init_var) to take a second length argument which
is already accepted by the underlying :generate method.