Add :forbid_id? method for case insensitivity
Exclusions passed to the RIG are likely to be variable names in
target script languages, which may or may not be case sensitive.
In order to avoid collision in case insensitive runtimes, convert
the check for presence in the forbidden set from direct comaprison
to case insensitive regular expression matching of the entire id.